% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{print hello with the name input}
\usage{
hello(InputName)
}
\arguments{
\item{InputName}{Name to be printed after}
}
\value{
prints the sentence
}
\description{
print hello with the name input
}
\examples{
hello("Rick")
}
